+++ py.bolt.snaek/CHANGELOG.md
@@ -7,6 +7,7 @@ versioned with [calendar versioning][calver].
 
 ## Changes
 
+- feat: thread responses to any message involving this scalable app 2024-02-25
 - build: prepare a development environment using a simple nix flake 2024-02-24
 - feat: respond to simple greetings in a slithery way via a CLI run 2024-02-24